requirements.txt
Django~=4.2.2
djangorestframework~=3.14.0
gTTS~=2.3.2
requests~=2.31.0
langdetect==1.0.9